Struct async_dup::Mutex

pub struct Mutex<T>(_);

A mutex that implements async I/O traits.

This is a blocking mutex that adds the following impls:

impl<T> AsyncRead for Mutex<T> where T: AsyncRead + Unpin {}

impl<T> AsyncRead for &Mutex<T> where T: AsyncRead + Unpin {}

impl<T> AsyncWrite for Mutex<T> where T: AsyncWrite + Unpin {}

impl<T> AsyncWrite for &Mutex<T> where T: AsyncWrite + Unpin {}

impl<T> AsyncSeek for Mutex<T> where T: AsyncSeek + Unpin {}

impl<T> AsyncSeek for &Mutex<T> where T: AsyncSeek + Unpin {}

Implementations

impl<T> Mutex<T>

pub fn new(data: T) -> Mutex<T>

Creates a new mutex.

Examples

use async_dup::Mutex;

let mutex = Mutex::new(10);

pub fn lock(&self) -> MutexGuard<'_, T>

Acquires the mutex, blocking the current thread until it is able to do so.

Returns a guard that releases the mutex when dropped.

Examples

use async_dup::Mutex;

let mutex = Mutex::new(10);
let guard = mutex.lock();
assert_eq!(*guard, 10);

pub fn try_lock(&self) -> Option<MutexGuard<'_, T>>

Attempts to acquire the mutex.

If the mutex could not be acquired at this time, then None is returned. Otherwise, a guard is returned that releases the mutex when dropped.

Examples

use async_dup::Mutex;

let mutex = Mutex::new(10);
if let Some(guard) = mutex.try_lock() {
    assert_eq!(*guard, 10);
}

pub fn into_inner(self) -> T

Consumes the mutex, returning the underlying data.

Examples

use async_dup::Mutex;

let mutex = Mutex::new(10);
assert_eq!(mutex.into_inner(), 10);

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the mutex mutably, no actual locking takes place – the mutable borrow statically guarantees the mutex is not already acquired.

Examples

use async_dup::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut() = 10;
assert_eq!(*mutex.lock(), 10);